requirements.txt
python-gitlab
pytz